+++ packages/adminui-vue/adminui-vue-types.js
@@ -48,6 +48,7 @@
  * @property {boolean} [isHtml=false] If set true will render content as v-html directive
  * @property {UTableColumnFormat} [format] Function what returns a formatted cell value to be inserted as cell innerHTML. Ignored in case slot is defined for column.
  * @property {UTableColumnFormat} [exportFormat] Function what returns a formatted cell value to be used during exports. If omitted - value will be used as is.
+ * @property {UTableColumnFormatXls} [exportFormatXlsColumn] Function that returns excel format depends on the column configuration
  * @property {number} [padding=16] column cells padding.
  * @property {object<string, UTableColumnFilter>} [filters={}] Filters templates
  */
@@ -67,6 +68,14 @@
  * @param {UTableColumn} column
  * @param {object} row
  * @returns {String}
+ */
+
+/**
+ * It is comfortable to return some value of the `XLSXStyle.predefinedFormats` object since these styles are already registered by XLSX
+ *
+ * @callback UTableColumnFormatXls
+ * @param {UTableColumn} column
+ * @returns {string|number}
  */
 
 /**