+++ packages/adminui-pub/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.20.12] - 2021-05-05
+### Fixed
  - Sidebar: prevent show en error in case user click on empty folder or on shortcut with empty command code,
    write error into console instead [LDOC-1434].